Miscellaneous updates remove sparkfun name from IDE option menu add Arduboy2 library version 6.0.0 changes add SH1106 I2C display support to Arduboy and Arduboy2 library Thanks to @evgenykzz2 add experimental code for extra X and Y buttons to Arduboy2 library add alternate wiring support to Arduboy2beep Thanks to @menehune23 update Arduboy2 library HardwareTest sketch add link to Cathy3K Repo in bootloader folder add contrast option to IDE option menu

+++ board-package-source/libraries/ArduboyFX/src/ArduboyFX.cpp
@@ -669,7 +669,7 @@ void FX::readDataArray(uint24_t address, uint8_t index, uint8_t offset, uint8_t 
 }
 
 
-uint16_t FX::readIndexedUInt8(uint24_t address, uint8_t index)
+uint8_t FX::readIndexedUInt8(uint24_t address, uint8_t index)
 {
   seekDataArray(address, index, 0, sizeof(uint8_t));
   return readEnd();
